Add separate Scan command This required to also create a Generate subcommand for the existing functionality.
